requirements.txt
# Web Framework & Server
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.3
pydantic-settings==2.2.1

# AI & Image Processing
google-genai==0.3.0
Pillow==10.2.0
pillow-avif-plugin==1.5.5

# Utilities
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
googlesearch-python==1.2.3